fix: normalize package names to lowercase and fix code style issues
- Rename all package directories to lowercase (config, controllers, models, etc.)
- Add javadoc to empty constructor in Event class
- Add private constructor to ProyectoietiApplication
- Replace tabs with spaces in ProyectoietiApplication.java
- Update test classes with new package imports
