--- src/main/java/com/ieti/proyectoieti/Config/OpenApiConfig.java
@@ -1,27 +1,0 @@
-package com.ieti.proyectoieti.Config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.info.Contact;
-import io.swagger.v3.oas.models.info.License;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenApiConfig {
-
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("EventIA API")
-                        .version("1.0.0")
-                        .description("API for EventIA - Event Management and Shared Wallet System")
-                        .contact(new Contact()
-                                .name("API Support")
-                                .email("support@eventia.com"))
-                        .license(new License()
-                                .name("Apache 2.0")
-                                .url("http://springdoc.org")));
-    }
-}
