--- src/main/java/com/ieti/proyectoieti/Exceptions/GlobalExceptionHandler.java
@@ -1,24 +1,0 @@
-package com.ieti.proyectoieti.Exceptions;
-
-import java.util.Map;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<?> handleIllegalArgument(IllegalArgumentException ex) {
-        return ResponseEntity
-            .status(HttpStatus.BAD_REQUEST)
-            .body(Map.of(
-                "error", "Bad Request",
-                "message", ex.getMessage()
-            ));
-    }
-
-   
-}
