+++ src/main/java/com/ieti/proyectoieti/ProyectoietiApplication.java
@@ -6,9 +6,13 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class ProyectoietiApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ProyectoietiApplication.class, args);
-		System.out.println("EventIA is running...");
-	}
+  /** Private constructor to prevent instantiation. */
+  private ProyectoietiApplication() {
+    // Utility class pattern
+  }
 
+  public static void main(String[] args) {
+    SpringApplication.run(ProyectoietiApplication.class, args);
+    System.out.println("EventIA is running...");
+  }
 }
