--- src/main/java/com/ieti/proyectoieti/Repositories/EventRepository.java
@@ -1,15 +1,0 @@
-package com.ieti.proyectoieti.Repositories;
-
-import com.ieti.proyectoieti.Models.Event;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Repository
-public interface EventRepository extends MongoRepository<Event, String> {
-    List<Event> findByCategory(String category);
-    List<Event> findByDateAfter(LocalDate date);
-    List<Event> findByLocationContainingIgnoreCase(String location);
-}
