+++ src/main/java/com/ieti/proyectoieti/config/OAuth2Config.java
@@ -1,0 +1,41 @@
+package com.ieti.proyectoieti.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.oauth2.client.registration.ClientRegistration;
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
+import org.springframework.security.oauth2.client.registration.InMemoryClientRegistrationRepository;
+import org.springframework.security.oauth2.core.AuthorizationGrantType;
+import org.springframework.security.oauth2.core.ClientAuthenticationMethod;
+
+@Configuration
+public class OAuth2Config {
+
+  @Value("${spring.security.oauth2.client.registration.google.client-id}")
+  private String clientId;
+
+  @Value("${spring.security.oauth2.client.registration.google.client-secret}")
+  private String clientSecret;
+
+  @Bean
+  public ClientRegistrationRepository clientRegistrationRepository() {
+    return new InMemoryClientRegistrationRepository(this.googleClientRegistration());
+  }
+
+  private ClientRegistration googleClientRegistration() {
+    return ClientRegistration.withRegistrationId("google")
+        .clientId(clientId)
+        .clientSecret(clientSecret)
+        .clientAuthenticationMethod(ClientAuthenticationMethod.CLIENT_SECRET_BASIC)
+        .authorizationGrantType(AuthorizationGrantType.AUTHORIZATION_CODE)
+        .redirectUri("{baseUrl}/login/oauth2/code/{registrationId}")
+        .scope("openid", "profile", "email")
+        .authorizationUri("https://accounts.google.com/o/oauth2/v2/auth")
+        .tokenUri("https://www.googleapis.com/oauth2/v4/token")
+        .userInfoUri("https://www.googleapis.com/oauth2/v3/userinfo")
+        .userNameAttributeName("sub")
+        .clientName("Google")
+        .build();
+  }
+}
